refactor(cli): extract clean-environment + grafana-health modules
Move the --clean docker lifecycle out of e2e.ts into a deep CleanEnvironment
module (src/cli/utils/clean-environment.ts) that encapsulates the single
"did we start the stack?" bit. Ownership is claimed inside reset() before
`up -d`, and teardownIfOwned() keeps the flag private \u2014 so the mid-reset
teardown-ordering bug is now impossible by construction rather than patched at
the call sites.

Also split the Grafana health probe into src/cli/utils/grafana-health.ts,
shared by the CLI pre-flight check and the clean reset's readiness poll
(avoids a cycle back into e2e.ts).
